app/requirements.txt
pillow
paho-mqtt
numpy
opencv-python==4.1.2.30
